Piksel difference
This Python code includes a function that calculates the pixel difference between two different images. First, it checks the dimensions of the two given images and, if they are not, resizes them to equalize the dimensions. It then calculates the difference in pixels between the two images, returning the intensity of the difference as a number and printing it on the screen.