Edge Detection
This Python code contains a program that detects the edges of the image by loading the image named "segmentationn.png", converting it to grayscale, and then applying edge detection. Edge is detected on the gray image using the Canny method and the detected edges are displayed in a window called "Edges".