Virtual Keyboard
This Python code creates an interface that can be used like a keyboard using the computer's camera. It takes images from the camera and detects objects of certain color in the image. It determines the letters and characters on the keyboard according to the positions of the detected objects and uses the detected characters as keyboard input. It determines the letters and characters on the keyboard according to the positions of the detected objects and uses the detected characters as keyboard input. In this way, you can enter text by matching an object in your hand with certain letters and characters in front of the camera.