Mastermind
This Python code describes a Mastermind (Number Guessing) game created using the tkinter library. In the game, the user tries to guess the order of the 4 randomly determined balls. Predicted colors are visualized on the screen as those in the right places and those in the wrong places. The player has a maximum of 12 guesses and if he guesses the correct order, the message "you win" is displayed. If the correct result is not reached even after 12 guesses, the correct answer is announced.