collect a ball
This Python code allows to detect a toy ball by receiving a live image stream from the camera and track the ball to calculate the player's score. The player tries to hold the ball, which is a blue object in the camera's field of view. When he catches the ball, his score increases and he tries to reach the highest score in a given time. During the game, game information is displayed on the screen and when the game is over, a message is shown with the score. It is also possible to convert the game to a recorded video.