Card_game_simulation
This Python code simulates a card game. The player draws his own and the king's cards and calculates their total. The player can continue or stop drawing cards as their total approaches 21. Its purpose is to set the total value to be as close to or equal to 21 as possible. There is a rule of drawing cards until the total of the king reaches 17.